feat: usage times over 24 hour period
