+++ cli/internal/ui/dashboard/cards/referrers.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/charmbracelet/lipgloss"
-	"github.com/tom-draper/nginx-analytics/cli/internal/logger"
 	"github.com/tom-draper/nginx-analytics/cli/internal/logs/nginx"
 	"github.com/tom-draper/nginx-analytics/cli/internal/logs/period"
 	"github.com/tom-draper/nginx-analytics/cli/internal/ui/styles"
@@ -104,8 +103,6 @@ func (p *ReferrersCard) RenderContent(width, height int) string {
 		Foreground(lipgloss.Color("15")) // White/default text
 
 	var lines []string
-
-	logger.Log.Println("referrers", referrers)
 
 	// Calculate how many referrers we can actually display
 	maxDisplayReferrers := min(len(referrers), height)
